[CDEC-353] Refactor Pos.Core.Txp module set
