+++ core/src/Pos/Core/Txp/TxWitness.hs
@@ -1,0 +1,61 @@
+module Pos.Core.Txp.TxWitness
+       ( TxWitness
+       , TxInWitness (..)
+       , TxSigData (..)
+       , TxSig
+       )where
+
+import           Universum
+
+import qualified Data.Text.Buildable as Buildable
+import           Formatting (bprint, build, (%))
+import           Serokell.Util.Base16 (base16F)
+
+import           Pos.Core.Common (Script, addressHash)
+import           Pos.Crypto (Hash, PublicKey, RedeemPublicKey, RedeemSignature, Signature, hash,
+                             shortHashF)
+
+import           Pos.Core.Txp.Tx (Tx)
+
+-- | A witness is a proof that a transaction is allowed to spend the funds it
+-- spends (by providing signatures, redeeming scripts, etc). A separate proof
+-- is provided for each input.
+type TxWitness = Vector TxInWitness
+
+-- | A witness for a single input.
+data TxInWitness
+    = PkWitness { twKey :: !PublicKey
+                , twSig :: !TxSig }
+    | ScriptWitness { twValidator :: !Script
+                    , twRedeemer  :: !Script }
+    | RedeemWitness { twRedeemKey :: !RedeemPublicKey
+                    , twRedeemSig :: !(RedeemSignature TxSigData) }
+    | UnknownWitnessType !Word8 !ByteString
+    deriving (Eq, Show, Generic, Typeable)
+
+instance Hashable TxInWitness
+
+instance Buildable TxInWitness where
+    build (PkWitness key sig) =
+        bprint ("PkWitness: key = "%build%", key hash = "%shortHashF%
+                ", sig = "%build) key (addressHash key) sig
+    build (ScriptWitness val red) =
+        bprint ("ScriptWitness: "%
+                "validator hash = "%shortHashF%", "%
+                "redeemer hash = "%shortHashF) (hash val) (hash red)
+    build (RedeemWitness key sig) =
+        bprint ("PkWitness: key = "%build%", sig = "%build) key sig
+    build (UnknownWitnessType t bs) =
+        bprint ("UnknownWitnessType "%build%" "%base16F) t bs
+
+instance NFData TxInWitness
+
+-- | Data that is being signed when creating a TxSig.
+data TxSigData = TxSigData
+    { -- | Transaction that we're signing
+      txSigTxHash      :: !(Hash Tx)
+    }
+    deriving (Eq, Show, Generic, Typeable)
+
+-- | 'Signature' of addrId.
+type TxSig = Signature TxSigData
